[PECOFF] Create the import library file.
If the linker is instructed to create a DLL, it will also create an import
library (.lib file) to describe the symbols exported by the DLL. This patch is
to create the import library file.

There is a convenient command "lib.exe" which can create an import library
from a module definition file (.def file). The command is used in this patch.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@197801 91177308-0d34-0410-b5e6-96231b3b80d8