[PECOFF] Create noname or data symbols as such.
If a symbol in an import library is marked as "data", the linker will not
create a jump table entry for the symbol, since jump table makes sense only
for a symbol pointing to a function.

I don't think NONAME attribute has a meaning when creating an import library.
The attribute is emitted for debugging purpose.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@197803 91177308-0d34-0410-b5e6-96231b3b80d8