Add make_dynamic_error_code().
This is a short-term fix to allow lld Readers to return error messages
with dynamic content.  

The long term fix will be to enhance ErrorOr<> to work with errors other
than error_code.  Or to change the interface to Readers to pass down a
diagnostics object through which all error messages are written.



git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@209681 91177308-0d34-0410-b5e6-96231b3b80d8