[mach-o]: make sure custom sort method is irreflexive.
The previous function returned true for "s < s", which could completely mess up
the sorting of symbols within a section.

Unfortunately, I don't think there's a robust way to write a test for this.
Anything I come up with will be making assumptions about the particular
implementation of std::sort.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@211704 91177308-0d34-0410-b5e6-96231b3b80d8