[mips] Remove .end and .end from test/elf/Mips/hilo16-3.test
The commit of the .ent/.end implementation will change the result of the
relocation evaluations (because of a new section and additional relocations)
which will lead to a failure if the .ent/.end directives are present in this
test.
We don't really need the .ent/.end directives in this test so let's just remove
them to preserve the current output.



git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@215534 91177308-0d34-0410-b5e6-96231b3b80d8