[ELF][Driver] Produce a proper error when file is not found
When a file is not found, produce a proper error message. The previous error
message produced a file format error, which made me wonder for a while why
there is a file format error, but essentially the file was not found.

This fixes the problem by producing a proper error message.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@217359 91177308-0d34-0410-b5e6-96231b3b80d8