[mach-o] Add support for -pie and -no_pie
There is a bit (MH_PIE) in the flags field of the mach_header which tells
the kernel is a program was built position independent (for ASLR).  The linker
automatically attempts to build programs PIE if they are built for a recent
OS version.  But the -pie and -no_pie options override that default behavior.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@217408 91177308-0d34-0410-b5e6-96231b3b80d8