Remove dead code.
Because we use cast<> at the beginning of this function, it will
abort there if a given atom is not a DefinedAtom.

In the switch statement, we checked if a given atom is a DefinedAtom
again by evaluating definition() == Atom::definitionRegular.
This was always true. So we can remove the outer switch statement.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@219724 91177308-0d34-0410-b5e6-96231b3b80d8