[macho] Create references from __eh_frame FDEs to their function.
We'll also need references back to the CIE eventually, but for now making sure
we can work out what an FDE is referring to is enough.

The actual kind of reference needs to be different between architectures,
probably because of MachO's chronic shortage of relocation types but I don't
really want to know in case I find out something that distresses me even more.

rdar://problem/18208653

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@219824 91177308-0d34-0410-b5e6-96231b3b80d8