[mach-o] enhance arm64 reloc parser to support local pointer relocations
The arm64 assembler almost always uses r_extern=1 relocations in which the
r_symbolnum field is the index of the symbol the relocation references.  But
sometimes it will set r_extern=0 in which case the linker needs to read the
content of the reloction to determine the target.

Add test case that the r_extern=0 relocation round trips.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@222200 91177308-0d34-0410-b5e6-96231b3b80d8