Remove InputGraph::registerObserver.
PECOFF was the only user of the API, and the reason why we created
the API is because, although the driver creates a list of input files,
it has no knowledge on what files are being created. It was because
everything was hidden behind the InputGraph abstraction.

Now the driver knows what that's doing. We no longer need this
indirection to get the file list being processed.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@225767 91177308-0d34-0410-b5e6-96231b3b80d8