Fix --start-group/end-group.
We used to manage the state whether we are in a group or not
using a counter. The counter is incremented by one if we jump from
end-group to start-group, and decremented by one if we don't.
The counter was assumed to be either zero or one, but obviously it
could be negative (if there's a group which is not repeated at all).
This is a fix for that issue.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@226632 91177308-0d34-0410-b5e6-96231b3b80d8