Driver: Replace switch cases with ifs.
We used to do like this instead of putting all command line processing
code within one gigantic switch statement. It is converted to a switch
in r188958, which introduced InputGraph.

In this patch I roll that change back. Now all "break"s are removed,
and the nesting is one level shallow.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@228646 91177308-0d34-0410-b5e6-96231b3b80d8