[ELF] Create a map from Reference to Symbol.
In LLD's model, symbol is a property of the node (atom) and not a property of
edge (reference). Prior to this patch, we stored the symbol in the reference.
From post-commit comments, it seemed better to create a map from the reference
to the symbol instead and use this mapping wherever desired.

Address comments from Ruiu/Simon Atanasyan.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@230273 91177308-0d34-0410-b5e6-96231b3b80d8