COFF: Fix a bug of __imp_ symbol.
The change I made in r240620 was not correct. If a symbol foo is
defined, and if you use __imp_foo, __imp_foo symbol is automatically
defined as a pointer (not just an alias) to foo.

Now that we need to create a chunk for automatically-created symbols.
I defined LocalImportChunk class for them.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@240622 91177308-0d34-0410-b5e6-96231b3b80d8