ELF: Drop the Chunk base class.
With OutputSection being a virtual interface, each concrete OutputSection
handles only one type of chunk and we don't need a base Chunk class.

So for we have a class that handles input sections and one that handles
the string table, but this extends naturally for other outputs (symbol table,
merging of SHF_MERGE sections, etc.).

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@244972 91177308-0d34-0410-b5e6-96231b3b80d8