[ELF2] Make the system page size a target-dependent property
The required page alignment is different on different targets. On PowerPC, for
example, we need 64K pages (the loader won't set different permissions on a
finer granularity than that). I've set the existing targets to what I believe
to be the correct values, and have updated the regression tests accordingly.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@249760 91177308-0d34-0410-b5e6-96231b3b80d8